refactor: split organization.go monolith into domain-specific files
Split lib/organization.go (984 lines, 46 functions) into 8 files:
- organization.go: core org CRUD, members, repos, collaborators (10 funcs)
- org_team.go: team management, members, permissions, invites (13 funcs)
- org_robot.go: robot accounts, permissions, federation (11 funcs)
- proxy_cache.go: proxy cache configuration (3 funcs)
- quota.go: quota management (4 funcs)
- auto_prune.go: auto-prune policies (5 funcs)
- application.go: OAuth applications (6 funcs)
- org_marketplace.go: marketplace subscriptions (4 funcs)

Closes #132
